added uploadmiddleware, updates codes for image upload but still in progress

+++ backend/controllers/productController.js
@@ -1,19 +1,43 @@
 const Product = require("../Models/product");
+const { productImageParser } = require("../middleware/uploadMiddleware");
 
-const createProduct = async(req, res) =>{
+// Use the productImageParser middleware for uploading product images
+const uploadProductImage = productImageParser.single("image");
+
+const createProduct = async (req, res) => {
   try {
+    // Set the user ID from the authentication token
     req.body.userId = req.user.userId;
-    const newProduct = await Product.create(req.body)
-    res.status(201).send({ msg: "Product created successfully", newProduct });
+
+    // Use the uploadProductImage middleware for handling image upload
+    uploadProductImage(req, res, async (err) => {
+      if (err) {
+        return res.status(500).send({ msg: "Failed to upload product image", error: err });
+      }
+
+      // if a new image is uploaded save the Cloudinary image URL to the product
+      if (req.file) {
+        req.body.image = req.file.path;
+        req.body.imagePublicId = req.file.filename;
+      }
+
+      // Create a new product in the database and response
+      const newProduct = await Product.create(req.body);
+      res.status(201).send({ msg: "Product created successfully", newProduct });
+    });
   } catch (error) {
-    console.log("error creation", error)
+    console.log("error creation", error);
     res.status(500).send({ msg: "Unable to create product", error });
   }
 }
 
 const getAllProducts = async(req, res) =>{
   try {
+
+    // Fetch all products from the database
     const products = await Product.find();
+
+    // Send the list of products as a response
     res.send(products)
   } catch (error) {
     res.status(500).send({ error: "An error occurred while getting products" });
@@ -22,9 +46,31 @@ const getAllProducts = async(req, res) =>{
 
 const updateProduct = async(req, res) =>{
   try {
+
+    // Set the user ID from the authentication token
     req.body.userId = req.user.userId;
-    await Product.findByIdAndUpdate({_id: req.params.id}, req.body)
-    res.send({ msg: "Product updated successfully"});
+
+    // if a new image uploaded save the Cloudinary image URL an ID to the product
+    if (req.file) {
+      req.body.image = req.file.path;
+      req.body.imagePublicId = req.file.filename; 
+    }
+
+    // Find the product by ID
+    const product = await Product.findById(req.params.id);
+
+    // Check if the authenticated user owns the product
+    if (product.userId !== req.user.userId) {
+      return res.status(403).send({ msg: "Unauthorized to update this product" });
+    }
+
+     // Find by ID and update only the specified fields using $set
+    await Product.findByIdAndUpdate(
+      req.params.id,
+      { $set: req.body },
+      { new: true } //true returns the updated product object
+    );
+    res.status(200).send({ msg: "Product updated successfully", updatedProduct });
   } catch (error) {
     res.status(500).send({ msg: "Unable to update product", error });
   }
@@ -32,6 +78,16 @@ const updateProduct = async(req, res) =>{
 
 const deleteProduct = async(req, res) =>{
   try {
+
+    // Find the product by ID
+    const product = await Product.findById(req.params.id);
+    
+    // Check if the authenticated user owns the product
+    if (product.userId !== req.user.userId) {
+      return res.status(403).send({ msg: "Unauthorized to delete this product" });
+    }
+
+    // Find the product by ID and delete it
     await Product.findByIdAndDelete({_id: req.params.id})
     res.send({ msg: "Deleted successfully" });
   } catch (error) {
@@ -39,8 +95,10 @@ const deleteProduct = async(req, res) =>{
   }
 }
 
+// get all products associated with a specific user
 const getAllUserProducts = async(req, res) =>{
   try {
+    // Find all products of a user based on their user ID
     const userProducts = await Product.find({ userId: req.user.userId })
     res.send(userProducts)
   } catch (error) {
@@ -53,5 +111,5 @@ module.exports = {
   getAllProducts,
   updateProduct,
   deleteProduct,
-  getAllUserProducts
+  getAllUserProducts,
 }
